pyproject.toml
[project]
name = "langgraph-deep-research"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.118.0",
    "ipykernel>=6.30.1",
    "langchain>=0.3.27",
    "langchain-groq>=0.3.8",
    "langgraph>=0.6.8",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.37.0",
]
